added api endpoint on backend to change affiliate bank account number

+++ src/backend/users/models.py
@@ -172,7 +172,7 @@ class AffiliateManager(models.Manager):
 
 class Affiliate(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    bank_account_number = models.IntegerField(null=True)
+    bank_account_number = models.PositiveBigIntegerField(null=True)
     #payment_email = models.EmailField()
     #amount_earned = models.DecimalField(max_digits=10, decimal_places=2)
     #next_payout = models.DecimalField(max_digits=10, decimal_places=2)
